fix: resolve remaining test file linting issues
- Fixed all remaining errcheck issues in o2-client test files
- Fixed all remaining errcheck issues in tn test files
- Fixed type mismatch in w.Write calls
- Added proper error handling for defer statements

Now all critical modules pass golangci-lint:
- o2-client: 0 issues ✅
- tn: Only non-critical test warnings remain
