+++ ran-dms/cmd/dms/main.go
@@ -1,0 +1,573 @@
+package main
+
+import (
+	"context"
+	"flag"
+	"fmt"
+	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
+	"golang.org/x/time/rate"
+
+	"github.com/thc1006/O-RAN-Intent-MANO-for-Network-Slicing/pkg/security"
+)
+
+// Version information
+var (
+	version = "v1.0.0"
+	build   = "unknown"
+)
+
+// Config holds the application configuration
+type Config struct {
+	Server struct {
+		Port       int    `mapstructure:"port"`
+		TLSPort    int    `mapstructure:"tls_port"`
+		TLSEnabled bool   `mapstructure:"tls_enabled"`
+		CertFile   string `mapstructure:"cert_file"`
+		KeyFile    string `mapstructure:"key_file"`
+	} `mapstructure:"server"`
+
+	Logging struct {
+		Level  string `mapstructure:"level"`
+		Format string `mapstructure:"format"`
+	} `mapstructure:"logging"`
+
+	Metrics struct {
+		Enabled bool `mapstructure:"enabled"`
+		Port    int  `mapstructure:"port"`
+	} `mapstructure:"metrics"`
+
+	RAN struct {
+		UpdateInterval time.Duration `mapstructure:"update_interval"`
+		MaxCells       int          `mapstructure:"max_cells"`
+		MaxSlices      int          `mapstructure:"max_slices"`
+	} `mapstructure:"ran"`
+
+	O2 struct {
+		Enabled  bool   `mapstructure:"enabled"`
+		Endpoint string `mapstructure:"endpoint"`
+		Timeout  time.Duration `mapstructure:"timeout"`
+	} `mapstructure:"o2"`
+}
+
+// Metrics
+var (
+	requestsTotal = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "ran_dms_requests_total",
+			Help: "Total number of requests to RAN DMS",
+		},
+		[]string{"method", "endpoint", "status"},
+	)
+
+	cellsActive = prometheus.NewGauge(
+		prometheus.GaugeOpts{
+			Name: "ran_dms_cells_active",
+			Help: "Number of active RAN cells",
+		},
+	)
+
+	slicesActive = prometheus.NewGauge(
+		prometheus.GaugeOpts{
+			Name: "ran_dms_slices_active",
+			Help: "Number of active RAN slices",
+		},
+	)
+
+	operationDuration = prometheus.NewHistogramVec(
+		prometheus.HistogramOpts{
+			Name:    "ran_dms_operation_duration_seconds",
+			Help:    "Duration of RAN DMS operations",
+			Buckets: prometheus.DefBuckets,
+		},
+		[]string{"operation"},
+	)
+)
+
+func init() {
+	// Register metrics
+	prometheus.MustRegister(requestsTotal)
+	prometheus.MustRegister(cellsActive)
+	prometheus.MustRegister(slicesActive)
+	prometheus.MustRegister(operationDuration)
+}
+
+func main() {
+	// Parse flags
+	configFile := flag.String("config", "/config/ran-dms.yaml", "Path to configuration file")
+	showVersion := flag.Bool("version", false, "Show version information")
+	flag.Parse()
+
+	if *showVersion {
+		fmt.Printf("RAN DMS %s (build: %s)\n", version, build)
+		os.Exit(0)
+	}
+
+	// Load configuration
+	config := loadConfig(*configFile)
+
+	// Setup logging
+	setupLogging(config.Logging.Level, config.Logging.Format)
+
+	log.WithFields(log.Fields{
+		"version": version,
+		"build":   build,
+	}).Info("Starting RAN DMS")
+
+	// Create router
+	router := setupRouter(config)
+
+	// Start metrics server if enabled
+	if config.Metrics.Enabled {
+		go startMetricsServer(config.Metrics.Port)
+	}
+
+	// Create server
+	srv := &http.Server{
+		Addr:              fmt.Sprintf(":%d", config.Server.Port),
+		Handler:           router,
+		ReadHeaderTimeout: 10 * time.Second,  // Prevent Slowloris attacks
+		ReadTimeout:       15 * time.Second,  // Total time to read request
+		WriteTimeout:      15 * time.Second,  // Time to write response
+		IdleTimeout:       60 * time.Second,  // Keep-alive timeout
+		MaxHeaderBytes:    1 << 20,           // 1MB max header size
+	}
+
+	// Start server in goroutine
+	go func() {
+		log.Infof("Starting HTTP server on port %d", config.Server.Port)
+		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+			log.Fatalf("Failed to start server: %v", err)
+		}
+	}()
+
+	// Start TLS server if enabled
+	if config.Server.TLSEnabled {
+		tlsSrv := &http.Server{
+			Addr:              fmt.Sprintf(":%d", config.Server.TLSPort),
+			Handler:           router,
+			ReadHeaderTimeout: 10 * time.Second,  // Prevent Slowloris attacks
+			ReadTimeout:       15 * time.Second,  // Total time to read request
+			WriteTimeout:      15 * time.Second,  // Time to write response
+			IdleTimeout:       60 * time.Second,  // Keep-alive timeout
+			MaxHeaderBytes:    1 << 20,           // 1MB max header size
+		}
+
+		go func() {
+			log.Infof("Starting HTTPS server on port %d", config.Server.TLSPort)
+			if err := tlsSrv.ListenAndServeTLS(config.Server.CertFile, config.Server.KeyFile); err != nil && err != http.ErrServerClosed {
+				log.Fatalf("Failed to start TLS server: %v", err)
+			}
+		}()
+	}
+
+	// Wait for interrupt signal
+	quit := make(chan os.Signal, 1)
+	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
+	<-quit
+
+	log.Info("Shutting down RAN DMS...")
+
+	// Graceful shutdown with timeout
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	if err := srv.Shutdown(ctx); err != nil {
+		log.Errorf("Server forced to shutdown: %v", err)
+	}
+
+	log.Info("RAN DMS stopped")
+}
+
+func loadConfig(configFile string) *Config {
+	config := &Config{}
+
+	// Set defaults
+	viper.SetDefault("server.port", 8080)
+	viper.SetDefault("server.tls_port", 8443)
+	viper.SetDefault("server.tls_enabled", false)
+	viper.SetDefault("logging.level", "info")
+	viper.SetDefault("logging.format", "json")
+	viper.SetDefault("metrics.enabled", true)
+	viper.SetDefault("metrics.port", 9090)
+	viper.SetDefault("ran.update_interval", "30s")
+	viper.SetDefault("ran.max_cells", 100)
+	viper.SetDefault("ran.max_slices", 50)
+	viper.SetDefault("o2.enabled", false)
+	viper.SetDefault("o2.endpoint", "http://localhost:8090")
+	viper.SetDefault("o2.timeout", "30s")
+
+	// Read from environment
+	viper.SetEnvPrefix("RAN_DMS")
+	viper.AutomaticEnv()
+
+	// Read config file if exists
+	viper.SetConfigFile(configFile)
+	if err := viper.ReadInConfig(); err != nil {
+		if _, ok := err.(viper.ConfigFileNotFoundError); !ok {
+			log.Fatalf("Error reading config file: %v", err)
+		}
+		log.Warnf("Config file not found, using defaults and environment variables")
+	}
+
+	// Unmarshal config
+	if err := viper.Unmarshal(config); err != nil {
+		log.Fatalf("Unable to decode config: %v", err)
+	}
+
+	return config
+}
+
+func setupLogging(level, format string) {
+	// Set log level
+	logLevel, err := log.ParseLevel(level)
+	if err != nil {
+		logLevel = log.InfoLevel
+		log.Warnf("Invalid log level %s, using info", level)
+	}
+	log.SetLevel(logLevel)
+
+	// Set log format
+	if format == "json" {
+		log.SetFormatter(&log.JSONFormatter{})
+	} else {
+		log.SetFormatter(&log.TextFormatter{
+			FullTimestamp: true,
+		})
+	}
+
+	// Output to stdout
+	log.SetOutput(os.Stdout)
+}
+
+func setupRouter(config *Config) *gin.Engine {
+	// Set Gin mode based on log level
+	if config.Logging.Level == "debug" {
+		gin.SetMode(gin.DebugMode)
+	} else {
+		gin.SetMode(gin.ReleaseMode)
+	}
+
+	router := gin.New()
+
+	// Add middleware
+	router.Use(customRecoveryMiddleware())
+	router.Use(securityHeadersMiddleware())
+	router.Use(rateLimitingMiddleware())
+	router.Use(loggingMiddleware())
+	router.Use(metricsMiddleware())
+
+	// API routes
+	api := router.Group("/api/v1")
+	{
+		// Slice management
+		api.GET("/slices", getSlices)
+		api.GET("/slices/:id", getSlice)
+		api.POST("/slices", createSlice)
+		api.PUT("/slices/:id", updateSlice)
+		api.DELETE("/slices/:id", deleteSlice)
+
+		// Cell management
+		api.GET("/cells", getCells)
+		api.GET("/cells/:id", getCell)
+		api.POST("/cells", createCell)
+		api.PUT("/cells/:id", updateCell)
+		api.DELETE("/cells/:id", deleteCell)
+
+		// RAN configuration
+		api.GET("/config", getRANConfig)
+		api.PUT("/config", updateRANConfig)
+
+		// Status and monitoring
+		api.GET("/status", getRANStatus)
+		api.GET("/metrics", getRANMetrics)
+		api.GET("/alarms", getAlarms)
+
+		// O-RAN O2 interface
+		api.GET("/o2/deployments", getO2Deployments)
+		api.POST("/o2/deployments", createO2Deployment)
+	}
+
+	// Health check
+	router.GET("/health", healthCheck)
+	router.GET("/ready", readinessCheck)
+
+	// Version
+	router.GET("/version", getVersion)
+
+	return router
+}
+
+func startMetricsServer(port int) {
+	mux := http.NewServeMux()
+	mux.Handle("/metrics", promhttp.Handler())
+
+	server := &http.Server{
+		Addr:              fmt.Sprintf(":%d", port),
+		Handler:           mux,
+		ReadHeaderTimeout: 10 * time.Second,  // Prevent Slowloris attacks
+		ReadTimeout:       30 * time.Second,  // Total time to read request
+		WriteTimeout:      30 * time.Second,  // Time to write response
+		IdleTimeout:       120 * time.Second, // Keep-alive timeout
+	}
+
+	log.Infof("Starting metrics server on port %d", port)
+	if err := server.ListenAndServe(); err != nil {
+		log.Errorf("Failed to start metrics server: %v", err)
+	}
+}
+
+// Middleware
+
+// customRecoveryMiddleware provides enhanced error handling and logging
+func customRecoveryMiddleware() gin.HandlerFunc {
+	return gin.CustomRecoveryWithWriter(os.Stderr, func(c *gin.Context, recovered interface{}) {
+		if err, ok := recovered.(string); ok {
+			// Sanitize log inputs to prevent log injection
+			log.WithFields(log.Fields{
+				"error":      security.SanitizeForLog(err),
+				"path":       security.SanitizeForLog(c.Request.URL.Path),
+				"method":     security.SanitizeForLog(c.Request.Method),
+				"client_ip":  security.SanitizeForLog(c.ClientIP()),
+				"user_agent": security.SanitizeForLog(c.Request.UserAgent()),
+			}).Error("Panic recovered in RAN-DMS")
+		}
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"error": "Internal server error",
+			"code":  "INTERNAL_ERROR",
+		})
+	})
+}
+
+// sanitizeLogInput removes control characters and limits length to prevent log injection
+
+// securityHeadersMiddleware adds security headers to all responses
+func securityHeadersMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		// Security headers
+		c.Header("X-Content-Type-Options", "nosniff")
+		c.Header("X-Frame-Options", "DENY")
+		c.Header("X-XSS-Protection", "1; mode=block")
+		c.Header("Strict-Transport-Security", "max-age=31536000; includeSubDomains")
+		c.Header("Referrer-Policy", "strict-origin-when-cross-origin")
+		c.Header("Content-Security-Policy", "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data:; connect-src 'self'")
+		c.Header("Permissions-Policy", "geolocation=(), microphone=(), camera=()")
+
+		// Remove server information
+		c.Header("Server", "")
+
+		c.Next()
+	}
+}
+
+// rateLimitingMiddleware implements rate limiting
+func rateLimitingMiddleware() gin.HandlerFunc {
+	limiter := rate.NewLimiter(rate.Limit(100), 200) // 100 req/sec with burst of 200
+
+	return func(c *gin.Context) {
+		if !limiter.Allow() {
+			c.JSON(http.StatusTooManyRequests, gin.H{
+				"error": "Rate limit exceeded",
+				"retry_after": "1s",
+			})
+			c.Abort()
+			return
+		}
+		c.Next()
+	}
+}
+
+func loggingMiddleware() gin.HandlerFunc {
+	return gin.LoggerWithConfig(gin.LoggerConfig{
+		SkipPaths: []string{"/health", "/ready"},
+	})
+}
+
+func metricsMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		start := time.Now()
+
+		c.Next()
+
+		duration := time.Since(start)
+		status := fmt.Sprintf("%d", c.Writer.Status())
+
+		requestsTotal.WithLabelValues(c.Request.Method, c.Request.URL.Path, status).Inc()
+		operationDuration.WithLabelValues(c.Request.URL.Path).Observe(duration.Seconds())
+	}
+}
+
+// Handlers
+func healthCheck(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"status": "healthy",
+		"time":   time.Now().Unix(),
+	})
+}
+
+func readinessCheck(c *gin.Context) {
+	// TODO: Add actual readiness checks
+	c.JSON(http.StatusOK, gin.H{
+		"ready": true,
+		"time":  time.Now().Unix(),
+	})
+}
+
+func getVersion(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"version": version,
+		"build":   build,
+	})
+}
+
+func getSlices(c *gin.Context) {
+	// TODO: Implement slice retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"slices": []gin.H{},
+		"count":  0,
+	})
+}
+
+func getSlice(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement single slice retrieval
+	c.JSON(http.StatusNotFound, gin.H{
+		"error": fmt.Sprintf("Slice %s not found", id),
+	})
+}
+
+func createSlice(c *gin.Context) {
+	// TODO: Implement slice creation
+	c.JSON(http.StatusCreated, gin.H{
+		"message": "RAN slice creation initiated",
+		"id":      "ran-slice-001",
+	})
+}
+
+func updateSlice(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement slice update
+	c.JSON(http.StatusOK, gin.H{
+		"message": fmt.Sprintf("RAN slice %s updated", id),
+	})
+}
+
+func deleteSlice(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement slice deletion
+	c.JSON(http.StatusOK, gin.H{
+		"message": fmt.Sprintf("RAN slice %s deleted", id),
+	})
+}
+
+func getCells(c *gin.Context) {
+	// TODO: Implement cell retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"cells": []gin.H{},
+		"count": 0,
+	})
+}
+
+func getCell(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement single cell retrieval
+	c.JSON(http.StatusNotFound, gin.H{
+		"error": fmt.Sprintf("Cell %s not found", id),
+	})
+}
+
+func createCell(c *gin.Context) {
+	// TODO: Implement cell creation
+	c.JSON(http.StatusCreated, gin.H{
+		"message": "Cell creation initiated",
+		"id":      "cell-001",
+	})
+}
+
+func updateCell(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement cell update
+	c.JSON(http.StatusOK, gin.H{
+		"message": fmt.Sprintf("Cell %s updated", id),
+	})
+}
+
+func deleteCell(c *gin.Context) {
+	id := c.Param("id")
+	// TODO: Implement cell deletion
+	c.JSON(http.StatusOK, gin.H{
+		"message": fmt.Sprintf("Cell %s deleted", id),
+	})
+}
+
+func getRANConfig(c *gin.Context) {
+	// TODO: Implement RAN configuration retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"config": gin.H{
+			"max_cells":  100,
+			"max_slices": 50,
+			"version":    version,
+		},
+	})
+}
+
+func updateRANConfig(c *gin.Context) {
+	// TODO: Implement RAN configuration update
+	c.JSON(http.StatusOK, gin.H{
+		"message": "RAN configuration updated",
+	})
+}
+
+func getRANStatus(c *gin.Context) {
+	// TODO: Implement RAN status retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"status": "operational",
+		"cells":  0,
+		"slices": 0,
+		"uptime": time.Now().Unix(),
+	})
+}
+
+func getRANMetrics(c *gin.Context) {
+	// TODO: Implement RAN metrics retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"metrics": gin.H{
+			"throughput": "0 Mbps",
+			"latency":    "0 ms",
+			"prbUsage":   "0%",
+		},
+	})
+}
+
+func getAlarms(c *gin.Context) {
+	// TODO: Implement alarm retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"alarms": []gin.H{},
+		"count":  0,
+	})
+}
+
+func getO2Deployments(c *gin.Context) {
+	// TODO: Implement O2 deployment retrieval
+	c.JSON(http.StatusOK, gin.H{
+		"deployments": []gin.H{},
+		"count":       0,
+	})
+}
+
+func createO2Deployment(c *gin.Context) {
+	// TODO: Implement O2 deployment creation
+	c.JSON(http.StatusCreated, gin.H{
+		"message": "O2 deployment initiated",
+		"id":      "o2-deploy-001",
+	})
+}
